:recycle: [Refactor]: broadcast 컴포넌트에서 socket을 갖고 있고, 훅은 socket을 인자로 받도록 수정
