+++ apps/client/src/hooks/useMediasoupDevice.ts
@@ -1,9 +1,7 @@
 import * as mediasoupClient from 'mediasoup-client';
 import { RtpCapabilities, RtpCodecCapability, RtpHeaderExtension } from 'mediasoup-client/lib/RtpParameters';
-import { io, Socket } from 'socket.io-client';
 import { useEffect, useRef, useState } from 'react';
-
-const SERVER_URL = 'ws://localhost:3001';
+import { Socket } from 'socket.io-client';
 
 interface RtpCapabilitiesResponse {
   rtpCapabilities: {
@@ -12,30 +10,10 @@ interface RtpCapabilitiesResponse {
   };
 }
 
-export const useMediasoupTransport = () => {
-  const socketRef = useRef<Socket | null>(null);
+export const useMediasoupDevice = (socket: Socket) => {
   const rtpCapabilitiesRef = useRef<RtpCapabilities | null>(null);
   const deviceRef = useRef<mediasoupClient.Device | null>(null);
-
   const [deviceError, setDeviceError] = useState<Error | null>(null);
-
-  const getRtpCapabilities = async () => {
-    try {
-      const newSocket = io(SERVER_URL);
-      socketRef.current = newSocket;
-
-      newSocket.on('connect_error', error => {
-        setDeviceError(new Error(`WebSocket 연결 실패: ${error}`));
-      });
-
-      socketRef.current.emit('getRtpCapabilities', async (response: RtpCapabilitiesResponse) => {
-        rtpCapabilitiesRef.current = response.rtpCapabilities;
-        await createDevice();
-      });
-    } catch (err) {
-      setDeviceError(err instanceof Error ? err : new Error('RTP Capabilities 요청 실패'));
-    }
-  };
 
   const createDevice = async () => {
     try {
@@ -51,14 +29,15 @@ export const useMediasoupTransport = () => {
   };
 
   useEffect(() => {
-    getRtpCapabilities();
-
-    return () => {
-      if (socketRef.current?.connected) {
-        socketRef.current.disconnect();
-      }
-    };
-  }, []);
+    try {
+      socket.emit('getRtpCapabilities', async (response: RtpCapabilitiesResponse) => {
+        rtpCapabilitiesRef.current = response.rtpCapabilities;
+        await createDevice();
+      });
+    } catch (err) {
+      setDeviceError(err instanceof Error ? err : new Error('RTP Capabilities 요청 실패'));
+    }
+  }, [socket]);
 
   return {
     rtpCapabilities: rtpCapabilitiesRef.current,
